splitting up actions further
