+++ src/store/demeris-api/actions.ts
@@ -3,13 +3,12 @@ import { EmerisAPI, EmerisBase } from '@emeris/types';
 import axios, { AxiosResponse } from 'axios';
 import { ActionTree } from 'vuex';
 
-import usePool from '@/composables/usePool';
-import { GlobalActionTypes, GlobalGetterTypes, RootState } from '@/store';
+import { GlobalGetterTypes, RootState } from '@/store';
 import { Pool } from '@/types/actions';
 import { UserData } from '@/types/user';
-import { ActionParams, ChartPrices, LoadingState, SimpleSubscribable, Subscribable } from '@/types/util';
+import { ActionParams, SimpleSubscribable, Subscribable } from '@/types/util';
 import { validPools } from '@/utils/actionHandler';
-import { getOwnAddress, hashObject, keyHashfromAddress } from '@/utils/basic';
+import { getOwnAddress, hashObject } from '@/utils/basic';
 import EmerisError from '@/utils/EmerisError';
 import TendermintWS from '@/utils/TendermintWS';
 
@@ -17,7 +16,11 @@ import { RootStoreTyped } from '../';
 import { APIStore } from '.';
 import { ActionTypes } from './action-types';
 import { AirdropActions, AirdropActionsInterface } from './actions/airdrops';
+import { BalanceActions, BalanceActionsInterface } from './actions/balances';
 import { ChainActions, ChainActionsInterface } from './actions/chain';
+import { PriceActions, PriceActionsInterface } from './actions/prices';
+import { StakingActions, StakingActionsInterface } from './actions/staking';
+import { TransactionActions, TransactionActionsInterface } from './actions/transactions';
 import { MutationTypes } from './mutation-types';
 import { APIState } from './state';
 
@@ -88,72 +91,15 @@ export type Subscription<K extends keyof Actions> = {
 };
 export type Subscriptions = Subscription<keyof Actions>;
 
-export interface Actions extends ChainActionsInterface, AirdropActionsInterface {
-  //Transaction Logic Action types
-  [ActionTypes.GET_TX_STATUS](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.TicketReq>>,
-  ): Promise<EmerisAPI.TicketResponse>;
-  [ActionTypes.GET_NUMBERS_CHAIN](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.ChainAddrReq>>,
-  ): Promise<EmerisAPI.SeqNumber>;
-
-  //Balances Action types
-  [ActionTypes.GET_BALANCES](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.AddrReq>>,
-  ): Promise<EmerisAPI.Balances>;
-  [ActionTypes.GET_POOL_BALANCES](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.AddrReq>>,
-  ): Promise<EmerisAPI.Balances>;
-  [ActionTypes.GET_STAKING_BALANCES](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.AddrReq>>,
-  ): Promise<EmerisAPI.StakingBalances>;
-  [ActionTypes.GET_UNBONDING_DELEGATIONS](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.AddrReq>>,
-  ): Promise<EmerisAPI.UnbondingDelegations>;
-  [ActionTypes.GET_ALL_BALANCES](context: APIActionContext): Promise<EmerisAPI.Balances>;
-  [ActionTypes.GET_ALL_STAKING_BALANCES](context: APIActionContext): Promise<EmerisAPI.StakingBalances>;
-  [ActionTypes.GET_ALL_UNBONDING_DELEGATIONS](context: APIActionContext): Promise<EmerisAPI.UnbondingDelegations>;
-
+export interface Actions
+  extends ChainActionsInterface,
+    AirdropActionsInterface,
+    BalanceActionsInterface,
+    StakingActionsInterface,
+    PriceActionsInterface,
+    TransactionActionsInterface {
   //Pools Action types
   [ActionTypes.VALIDATE_POOLS](context: APIActionContext, pools: Pool[]): Promise<Pool[]>;
-
-  //Coingecko Action types
-  [ActionTypes.GET_COINGECKO_ID_BY_NAMES](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.TokenIdReq>>,
-  ): Promise<any>;
-
-  //Prices Action types
-  [ActionTypes.GET_PRICES](context: APIActionContext, payload: SimpleSubscribable): Promise<EmerisAPI.Prices>;
-  [ActionTypes.GET_TOKEN_PRICES](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.TokenPriceReq>>,
-  ): Promise<ChartPrices>;
-  [ActionTypes.RESET_TOKEN_PRICES](context: APIActionContext): void;
-
-  //Staking Action types
-  [ActionTypes.GET_INFLATION](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.ChainReq>>,
-  ): Promise<number>;
-  [ActionTypes.GET_STAKING_REWARDS](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.ChainReq>>,
-  ): Promise<EmerisAPI.StakingRewardsResponse>;
-  [ActionTypes.GET_UNSTAKING_PARAM](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.ChainReq>>,
-  ): Promise<EmerisAPI.StakingParams>;
-  [ActionTypes.GET_VALIDATORS](
-    context: APIActionContext,
-    payload: Subscribable<ActionParams<EmerisAPI.ChainReq>>,
-  ): Promise<EmerisAPI.Validator[]>;
 
   //Uncategorized Action types
   [ActionTypes.GET_VERIFIED_DENOMS](
@@ -177,169 +123,11 @@ export type GlobalActions = Namespaced<Actions, 'demerisAPI'>;
 export const actions: ActionTree<APIState, RootState> & Actions = {
   ...ChainActions,
   ...AirdropActions,
+  ...BalanceActions,
+  ...StakingActions,
+  ...PriceActions,
+  ...TransactionActions,
 
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  async [ActionTypes.GET_BALANCES]({ commit, dispatch, getters, state, rootGetters }, { subscribe = false, params }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    const reqHash = hashObject({ action: ActionTypes.GET_BALANCES, payload: { params } });
-
-    if (state._InProgess.get(reqHash)) {
-      await state._InProgess.get(reqHash);
-
-      return getters['getBalances'](params);
-    } else {
-      let resolver;
-      let rejecter;
-      const promise: Promise<void> = new Promise((resolve, reject) => {
-        resolver = resolve;
-        rejecter = reject;
-      });
-      commit(MutationTypes.SET_IN_PROGRESS, { hash: reqHash, promise });
-      try {
-        const response: AxiosResponse<EmerisAPI.BalancesResponse> = await axios.get(
-          getters['getEndpoint'] + '/account/' + params.address + '/balance',
-        );
-
-        if (response.data.balances) {
-          const tracesLoaded = [];
-          for (const balance of response.data.balances) {
-            if (
-              // balance.ibc holds resolved ibc denoms so we can show the name of the coin. this functions as a cache. so if it was loaded in the past we don't need to load it again.
-              Object.keys(balance.ibc).length != 0 &&
-              !getters['getVerifyTrace']({
-                chain_name: balance.on_chain,
-                hash: balance.ibc.hash,
-              })
-            ) {
-              tracesLoaded.push(
-                dispatch(ActionTypes.GET_VERIFY_TRACE, {
-                  subscribe: false,
-                  params: { chain_name: balance.on_chain, hash: balance.ibc.hash },
-                }),
-              );
-            }
-          }
-          await Promise.all(tracesLoaded);
-        }
-        commit(MutationTypes.SET_BALANCES, { params, value: response.data.balances });
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_BALANCES, payload: { params } });
-        }
-      } catch (e) {
-        commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-        rejecter(e);
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_BALANCES, payload: { params } });
-        }
-        throw new EmerisError('Demeris:GetBalances', 'Could not perform API query.');
-      }
-      commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-      resolver();
-      return getters['getBalances'](params);
-    }
-  },
-  async [ActionTypes.GET_POOL_BALANCES]({ commit, getters, state, rootGetters }, { subscribe = false, params }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    const reqHash = hashObject({ action: ActionTypes.GET_POOL_BALANCES, payload: { params } });
-
-    if (state._InProgess.get(reqHash)) {
-      await state._InProgess.get(reqHash);
-
-      return getters['getBalances'](params);
-    } else {
-      let resolver;
-      let rejecter;
-      const promise: Promise<void> = new Promise((resolve, reject) => {
-        resolver = resolve;
-        rejecter = reject;
-      });
-      commit(MutationTypes.SET_IN_PROGRESS, { hash: reqHash, promise });
-      try {
-        const response: AxiosResponse<EmerisAPI.BalancesResponse> = await axios.get(
-          getters['getEndpoint'] + '/account/' + params.address + '/balance',
-        );
-
-        commit(MutationTypes.SET_POOL_BALANCES, { params, value: response.data.balances });
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_POOL_BALANCES, payload: { params } });
-        }
-      } catch (e) {
-        commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-        rejecter(e);
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_POOL_BALANCES, payload: { params } });
-        }
-        throw new EmerisError('Demeris:GetPoolBalances', 'Could not perform API query.');
-      }
-      commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-      resolver();
-
-      return getters['getBalances'](params);
-    }
-  },
-  async [ActionTypes.GET_ALL_BALANCES]({ dispatch, getters, rootGetters }) {
-    try {
-      const keyHashes = rootGetters[GlobalGetterTypes.USER.getKeyhashes];
-
-      const balanceLoads = [];
-
-      const chains =
-        getters['getChains'] ??
-        (await dispatch(ActionTypes.GET_CHAINS, {
-          subscribe: false,
-        }));
-      for (const chain in chains) {
-        if (!chains[chain].primary_channel)
-          chains[chain] = await dispatch(ActionTypes.GET_CHAIN, {
-            subscribe: true,
-            params: {
-              chain_name: chain,
-            },
-          });
-      }
-      for (const keyHash of keyHashes) {
-        balanceLoads.push(dispatch(ActionTypes.GET_BALANCES, { subscribe: true, params: { address: keyHash } }));
-      }
-      await Promise.all(balanceLoads);
-      if (rootGetters[GlobalGetterTypes.USER.getBalancesFirstLoad]) {
-        dispatch(GlobalActionTypes.USER.BALANCES_LOADED, null, { root: true });
-      }
-    } catch (e) {
-      throw new EmerisError('Demeris:GetAllBalances', 'Could not perform API query.');
-    }
-    return getters['getAllBalances'];
-  },
-  async [ActionTypes.GET_ALL_STAKING_BALANCES]({ dispatch, getters, rootGetters }) {
-    try {
-      const keyHashes = rootGetters[GlobalGetterTypes.USER.getKeyhashes];
-
-      const stakingBalanceLoads = [];
-      for (const keyHash of keyHashes) {
-        stakingBalanceLoads.push(
-          dispatch(ActionTypes.GET_STAKING_BALANCES, { subscribe: true, params: { address: keyHash } }),
-        );
-      }
-      await Promise.all(stakingBalanceLoads);
-
-      if (rootGetters[GlobalGetterTypes.USER.getStakingBalancesFirstLoad]) {
-        dispatch(GlobalActionTypes.USER.STAKING_BALANCES_LOADED, null, { root: true });
-      }
-    } catch (e) {
-      throw new EmerisError('Demeris:GetAllStakingBalances', 'Could not perform API query.');
-    }
-    return getters['getAllStakingBalances'];
-  },
-  async [ActionTypes.GET_ALL_UNBONDING_DELEGATIONS]({ dispatch, getters, rootGetters }) {
-    try {
-      const keyHashes = rootGetters[GlobalGetterTypes.USER.getKeyhashes];
-      for (const keyHash of keyHashes) {
-        dispatch(ActionTypes.GET_UNBONDING_DELEGATIONS, { subscribe: true, params: { address: keyHash } });
-      }
-    } catch (e) {
-      throw new EmerisError('Demeris:GetAllUnbondingDelegations', 'Could not perform API query.');
-    }
-    return getters['getAllUnbondingDelegations'];
-  },
   async [ActionTypes.VALIDATE_POOLS]({ commit, getters }, pools) {
     try {
       const vp = await validPools(pools);
@@ -349,119 +137,6 @@ export const actions: ActionTree<APIState, RootState> & Actions = {
       throw new EmerisError('Demeris:ValidatePools', 'Could not perform pool validation.');
     }
     return getters['getAllValidPools'];
-  },
-  async [ActionTypes.GET_UNSTAKING_PARAM](
-    { commit, getters },
-    { subscribe: _subscribe, params: { chain_name } },
-  ): Promise<EmerisAPI.StakingParams> {
-    try {
-      const {
-        data: { params: unstakingParam },
-      } = await axios.get(`${getters['getEndpoint']}/chain/${chain_name}/staking/params`);
-      commit(MutationTypes.SET_UNSTAKING_PARAM, { params: { chain_name }, value: unstakingParam });
-      return getters['getUnstakingParam']({ chain_name });
-    } catch {
-      throw new EmerisError('Demeris:getUnstakingParam', 'Could not retrieve staking param.');
-    }
-  },
-  async [ActionTypes.GET_STAKING_BALANCES]({ commit, getters, state, rootGetters }, { subscribe = false, params }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    const reqHash = hashObject({ action: ActionTypes.GET_STAKING_BALANCES, payload: { params } });
-
-    if (state._InProgess.get(reqHash)) {
-      await state._InProgess.get(reqHash);
-
-      return getters['getStakingBalances'](params);
-    } else {
-      let resolver;
-      let rejecter;
-      const promise: Promise<void> = new Promise((resolve, reject) => {
-        resolver = resolve;
-        rejecter = reject;
-      });
-      commit(MutationTypes.SET_IN_PROGRESS, { hash: reqHash, promise });
-      try {
-        const response: AxiosResponse<EmerisAPI.StakingBalancesResponse> = await axios.get(
-          getters['getEndpoint'] + '/account/' + params.address + '/stakingbalances',
-        );
-        commit(MutationTypes.SET_STAKING_BALANCES, { params, value: response.data.staking_balances });
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_STAKING_BALANCES, payload: { params } });
-        }
-      } catch (e) {
-        commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-        rejecter(e);
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_STAKING_BALANCES, payload: { params } });
-        }
-        throw new EmerisError('Demeris:GetStakingBalances', 'Could not perform API query.');
-      }
-      commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-      resolver();
-
-      return getters['getStakingBalances'](params);
-    }
-  },
-  async [ActionTypes.GET_UNBONDING_DELEGATIONS](
-    { commit, getters, state, rootGetters },
-    { subscribe = false, params },
-  ) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    const reqHash = hashObject({ action: ActionTypes.GET_UNBONDING_DELEGATIONS, payload: { params } });
-
-    if (state._InProgess.get(reqHash)) {
-      await state._InProgess.get(reqHash);
-
-      return getters['getUnbondingDelegations'](params);
-    } else {
-      let resolver;
-      let rejecter;
-      const promise: Promise<void> = new Promise((resolve, reject) => {
-        resolver = resolve;
-        rejecter = reject;
-      });
-      commit(MutationTypes.SET_IN_PROGRESS, { hash: reqHash, promise });
-      try {
-        const response: AxiosResponse<EmerisAPI.UnbondingDelegationsResponse> = await axios.get(
-          getters['getEndpoint'] + '/account/' + params.address + '/unbondingdelegations',
-        );
-        commit(MutationTypes.SET_UNBONDING_DELEGATIONS, { params, value: response.data.unbonding_delegations });
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_UNBONDING_DELEGATIONS, payload: { params } });
-        }
-      } catch (e) {
-        commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-        rejecter(e);
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_UNBONDING_DELEGATIONS, payload: { params } });
-        }
-        throw new EmerisError('Demeris:GetUnbondingDelegations', 'Could not perform API query.');
-      }
-      commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-      resolver();
-
-      return getters['getUnbondingDelegations'](params);
-    }
-  },
-  /**
-   * Gets sequence and account number. Used when making a transaction.
-   * @param {string} chain_name - chain name
-   * @param {string} address - address
-   */
-  async [ActionTypes.GET_NUMBERS_CHAIN]({ commit, getters, rootGetters }, { subscribe = false, params }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    try {
-      const response: AxiosResponse<EmerisAPI.NumbersResponse> = await axios.get(
-        getters['getEndpoint'] + '/chain/' + params.chain_name + '/numbers/' + params.address,
-      );
-      commit(MutationTypes.SET_NUMBERS_CHAIN, { params, value: response.data.numbers });
-      if (subscribe) {
-        commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_NUMBERS_CHAIN, payload: { params } });
-      }
-    } catch (e) {
-      throw new EmerisError('Demeris:GetNumbersChain', 'Could not perform API query.');
-    }
-    return getters['getNumbersChain'](params);
   },
   async [ActionTypes.GET_VERIFIED_DENOMS]({ commit, getters, rootGetters }, { subscribe = false }) {
     axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
@@ -477,138 +152,6 @@ export const actions: ActionTree<APIState, RootState> & Actions = {
       throw new EmerisError('Demeris:GetVerifiedDenoms', 'Could not perform API query.');
     }
     return getters['getVerifiedDenoms'];
-  },
-  async [ActionTypes.GET_PRICES]({ commit, getters, rootGetters, state, dispatch }, { subscribe = false }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    const isCypress = !!window['Cypress'];
-    const reqHash = hashObject({ action: ActionTypes.GET_PRICES, payload: {} });
-
-    if (state._InProgess.get(reqHash)) {
-      await state._InProgess.get(reqHash);
-
-      return getters['getPrices'];
-    } else {
-      let resolver;
-      let rejecter;
-      const promise: Promise<void> = new Promise((resolve, reject) => {
-        resolver = resolve;
-        rejecter = reject;
-      });
-      commit(MutationTypes.SET_IN_PROGRESS, { hash: reqHash, promise });
-      try {
-        const response: AxiosResponse<EmerisAPI.PricesResponse> = await axios.get(
-          getters['getEndpoint'] + '/oracle/prices',
-        );
-        if (response.data?.data?.Tokens) {
-          if (isCypress) {
-            commit(MutationTypes.SET_PRICES, {
-              value: {
-                Fiats: response.data.data.Fiats,
-                Tokens: response.data.data.Tokens.map((x) => {
-                  return { ...x, Price: 1.1 };
-                }),
-              },
-            });
-          } else {
-            commit(MutationTypes.SET_PRICES, { value: response.data.data });
-          }
-          // Set initial prices so pool calculations can find them
-          await Promise.all(
-            getters['getVerifiedDenoms'].map(async (denom) => {
-              if (denom.name.startsWith('pool')) {
-                const pools = rootGetters['tendermint.liquidity.v1beta1/getLiquidityPools']().pools;
-
-                if (pools) {
-                  const pool = pools.find((pool) => pool.pool_coin_denom == denom.name);
-
-                  if (pool) {
-                    const { totalLiquidityPrice, totalSupply, initPromise } = usePool(pool.id);
-                    await initPromise;
-                    try {
-                      if (totalLiquidityPrice.value > 0) {
-                        const priceData = {
-                          Symbol: denom.ticker + 'USDT',
-                          Price: (totalLiquidityPrice.value * 10 ** 6) / totalSupply.value,
-                          Supply: totalSupply.value,
-                        };
-                        response.data.data.Tokens.push(priceData);
-                      }
-                    } catch (e) {}
-                  }
-                }
-              }
-            }),
-          );
-          if (isCypress) {
-            commit(MutationTypes.SET_PRICES, {
-              value: {
-                Fiats: response.data.data.Fiats,
-                Tokens: response.data.data.Tokens.map((x) => {
-                  return { ...x, Price: 1.1 };
-                }),
-              },
-            });
-          } else {
-            commit(MutationTypes.SET_PRICES, { value: response.data.data });
-          }
-          // Set prices incl. pool calculations
-
-          if (rootGetters[GlobalGetterTypes.USER.getPricesFirstLoad]) {
-            dispatch(GlobalActionTypes.USER.PRICES_LOADED, null, { root: true });
-          }
-        }
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_PRICES, payload: {} });
-        }
-      } catch (e) {
-        commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-        rejecter(e);
-        if (subscribe) {
-          commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_PRICES, payload: {} });
-        }
-        throw new EmerisError('Demeris:GetPrices', 'Could not perform API query.');
-      }
-      commit(MutationTypes.DELETE_IN_PROGRESS, reqHash);
-      resolver();
-      return getters['getPrices'];
-    }
-  },
-  async [ActionTypes.GET_TX_STATUS]({ commit, getters, rootGetters }, { subscribe = false, params }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    try {
-      const response: AxiosResponse<EmerisAPI.TicketResponse> = await axios.get(
-        getters['getEndpoint'] + '/tx/ticket/' + params.chain_name + '/' + params.ticket,
-      );
-      commit(MutationTypes.SET_TX_STATUS, { params, value: response.data });
-      if (subscribe) {
-        commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_TX_STATUS, payload: { params } });
-      }
-    } catch (e) {
-      console.error(e);
-      throw new EmerisError('Demeris:GetTXStatus', 'Could not perform API query.');
-    }
-    return getters['getTxStatus'](params);
-  },
-  async [ActionTypes.GET_TX_DEST_HASH]({ getters, rootGetters }, { from_chain, to_chain, txhash }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    try {
-      const response: AxiosResponse<EmerisAPI.DestinationTXResponse> = await axios.get(
-        `${getters['getEndpoint']}/tx/${from_chain}/${to_chain}/${txhash}`,
-      );
-      const data = response.data;
-
-      if (data.cause) {
-        throw new Error(data.cause);
-      }
-
-      if (!data.tx_hash) {
-        throw new Error('Failed to fetch destination hash');
-      }
-
-      return response.data;
-    } catch (e) {
-      throw new EmerisError('Demeris:GetTXDestHash', 'Could not perform API query.');
-    }
   },
 
   // Chain-specific endpoint actions
@@ -654,29 +197,6 @@ export const actions: ActionTree<APIState, RootState> & Actions = {
       return getters['getVerifyTrace'](params);
     }
   },
-  async [ActionTypes.GET_TOKEN_PRICES]({ commit, getters, rootGetters }, { subscribe = false, params }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    commit(MutationTypes.SET_TOKEN_PRICES_STATUS, {
-      value: params.showSkeleton ? LoadingState.LOADING : LoadingState.LOADED,
-    });
-    try {
-      const response: AxiosResponse<EmerisAPI.TokenPriceResponse> = await axios.get(
-        getters['getEndpoint'] + `/oracle/chart/${params.token_id}?days=${params.days}&vs_currency=${params.currency}`,
-      );
-      commit(MutationTypes.SET_TOKEN_PRICES, { value: response.data.data.prices });
-      commit(MutationTypes.SET_TOKEN_PRICES_STATUS, { value: LoadingState.LOADED });
-      if (subscribe) {
-        commit(MutationTypes.SUBSCRIBE, { action: ActionTypes.GET_TOKEN_PRICES, payload: { params } });
-      }
-    } catch (e) {
-      commit(MutationTypes.SET_TOKEN_PRICES_STATUS, { value: LoadingState.ERROR });
-      throw new EmerisError('Demeris:getTokenPrices', 'Could not perform API query.');
-    }
-    return getters['getTokenPrices'];
-  },
-  [ActionTypes.RESET_TOKEN_PRICES]({ commit }) {
-    commit(MutationTypes.SET_TOKEN_PRICES, { value: [] });
-  },
   async [ActionTypes.GET_NEW_BLOCK]({ getters }, { chain_name }) {
     return new Promise(async (resolve, reject) => {
       const timeout = 30000;
@@ -696,84 +216,6 @@ export const actions: ActionTree<APIState, RootState> & Actions = {
 
       setTimeout(reject, timeout);
     });
-  },
-  async [ActionTypes.TRACE_TX_RESPONSE]({ getters }, { txhash, chain_name }) {
-    return new Promise(async (resolve, reject) => {
-      const timeout = 60000;
-      const wsUrl = `${getters['getWebSocketEndpoint']}/chain/${chain_name}/rpc/websocket`;
-
-      const wss = new TendermintWS({ server: wsUrl, timeout: 5000, autoReconnect: false });
-      const txHash64 = Buffer.from(txhash, 'hex').toString('base64');
-      const subscribeQuery = `tm.event = 'Tx' AND tx.hash = '${txhash}'`;
-
-      let done = false;
-
-      const getTx = async () => {
-        const result = await wss.call('tx', [txHash64, false]).catch(reject);
-        handleMessage(result);
-      };
-
-      const subscribeTx = () => {
-        wss.subscribe(
-          {
-            query: subscribeQuery,
-          },
-          handleMessage,
-        );
-      };
-
-      const handleOpen = () => {
-        getTx();
-        subscribeTx();
-      };
-
-      const handleMessage = async (data: Record<string, any>) => {
-        if (done) return;
-
-        if (data.error) {
-          // Not found
-          if (data.error.code === -32603) return;
-
-          done = true;
-          reject(new Error(data.error));
-        }
-
-        if (data.result?.data?.value?.TxResult) {
-          done = true;
-          resolve(data.result);
-        }
-
-        if (data?.result?.tx_result) {
-          done = true;
-          resolve(data.result);
-        }
-      };
-
-      await wss.connect().catch(reject);
-      handleOpen();
-
-      setTimeout(() => {
-        done = true;
-        reject(new Error('Could not find transaction response'));
-      }, timeout);
-    });
-  },
-
-  async [ActionTypes.GET_TX_FROM_RPC]({ getters }, { txhash, chain_name }) {
-    const rpcUrl = `${getters['getEndpoint']}/chain/${chain_name}/rpc`;
-
-    if (!rpcUrl) {
-      throw new Error(`${chain_name} RPC endpoint not found`);
-    }
-
-    try {
-      delete axios.defaults.headers.get['X-Correlation-Id'];
-      const { data } = await axios.get(`${rpcUrl}/tx?hash=0x${txhash}`);
-
-      return data?.result;
-    } catch (e) {
-      throw new Error('Could not find transaction response from RPC');
-    }
   },
 
   async [ActionTypes.GET_END_BLOCK_EVENTS]({ getters, rootGetters }, { height, stepType }: DemerisTxResultParams) {
@@ -832,43 +274,6 @@ export const actions: ActionTree<APIState, RootState> & Actions = {
     }
   },
 
-  async [ActionTypes.GET_VALIDATORS]({ getters, rootGetters }, { subscribe: _subscribe, params: { chain_name } }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    try {
-      const response: AxiosResponse<EmerisAPI.ValidatorsResponse> = await axios.get(
-        getters['getEndpoint'] + '/chain/' + chain_name + '/validators',
-      );
-      return response.data?.validators;
-    } catch (e) {
-      throw new EmerisError('Demeris:GET_VALIDATORS', `Could not get ${chain_name} validators.` + e.message);
-    }
-  },
-
-  async [ActionTypes.GET_INFLATION]({ getters, rootGetters }, { subscribe: _subscribe, params: { chain_name } }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    try {
-      const response: AxiosResponse<EmerisAPI.InflationResponse> = await axios.get(
-        getters['getEndpoint'] + '/chain/' + chain_name + '/mint/inflation',
-      );
-      return Number(response.data?.inflation);
-    } catch (e) {
-      throw new EmerisError('Demeris:GET_INFLATION', `Could not get ${chain_name} inflation.` + e.message);
-    }
-  },
-
-  async [ActionTypes.GET_STAKING_REWARDS]({ getters, rootGetters }, { subscribe: _subscribe, params: { chain_name } }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    try {
-      const address = keyHashfromAddress(await getOwnAddress({ chain_name }));
-      const response: AxiosResponse<EmerisAPI.StakingRewardsResponse> = await axios.get(
-        getters['getEndpoint'] + '/account/' + address + '/delegatorrewards/' + chain_name,
-      );
-      return response.data;
-    } catch (e) {
-      throw new EmerisError('Demeris:GET_REWARDS', `Could not get ${chain_name} rewards.` + e.message);
-    }
-  },
-
   // Internal module actions
 
   [ActionTypes.INIT](
@@ -912,24 +317,5 @@ export const actions: ActionTree<APIState, RootState> & Actions = {
   },
   [ActionTypes.UNSUBSCRIBE]({ commit }, subscription) {
     commit(MutationTypes.UNSUBSCRIBE, subscription);
-  },
-
-  // Coingecko Actions
-  async [ActionTypes.GET_COINGECKO_ID_BY_NAMES]({ commit, getters, rootGetters }, { params }) {
-    axios.defaults.headers.get['X-Correlation-Id'] = rootGetters[GlobalGetterTypes.USER.getCorrelationId];
-    commit(MutationTypes.SET_COINGECKO_ID_STATUS, {
-      value: LoadingState.LOADING,
-    });
-    try {
-      const response: AxiosResponse<EmerisAPI.TokenIdResponse> = await axios.get(
-        getters['getEndpoint'] + `/oracle/geckoid?names=${params.token}`,
-      );
-      commit(MutationTypes.SET_COINGECKO_ID, { params: params.token, value: response.data });
-      commit(MutationTypes.SET_COINGECKO_ID_STATUS, { value: LoadingState.LOADED });
-      return getters['getCoinGeckoId'];
-    } catch (e) {
-      commit(MutationTypes.SET_COINGECKO_ID_STATUS, { value: LoadingState.ERROR });
-      console.error('Demeris:getCoinGeckoId: Could not perform API query.');
-    }
   },
 };
